ARM: sun8i: tbs-a711: Add device node for AXP813 PMIC
The AXP813/AXP818 PMICs used with the A83T/H8 SoCs are actually 2 dies
in one package sharing the serial bus (I2C/RSB) pins. One die is the
actual PMIC. The other is an AC100 codec / RTC combo chip.

This patch adds a device node for the PMIC die.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>