media: sun6i-csi: Add support for Allwinner CSI
Allwinner V3s SoC has two CSI modules. CSI0 is used for MIPI interface
and CSI1 is used for parallel interface. This is not documented in
datasheet but it was found by testing and guessing.

This patch implements a v4l2 framework driver for it.

Currently, the driver only support the parallel interface. MIPI-CSI2,
ISP's support are not included in this patch.

This patch was cleaned up and extended to support A83T by Ondrej Jirman.

Signed-off-by: Yong Deng <yong.deng@magewell.com>
Signed-off-by: Ondrej Jirman <megous@megous.com>