drm/bridge/synopsys: dw-hdmi: don't clobber drvdata
dw_hdmi shouldn't set drvdata since some drivers might need to store
it's own data there. Rework dw_hdmi in a way to return struct dw_hdmi
instead to store it in drvdata. This way drivers are responsible to
store and pass structure when needed.

Idea for this commit was taken from the following patch:
8242ecbd597d ("drm/bridge/synopsys: stop clobbering drvdata")

Cc: p.zabel@pengutronix.de
Cc: narmstrong@baylibre.com
Cc: Laurent.pinchart@ideasonboard.com
Cc: hjc@rock-chips.com
Cc: heiko@sntech.de
Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>